+++ driver-core/src/main/java/com/datastax/driver/core/ControlConnection.java
@@ -74,6 +74,7 @@ class ControlConnection implements Host.StateListener {
         if (isShutdown)
             return;
 
+        // NB: at this stage, allHosts() only contains the initial contact points
         setNewConnection(reconnectInternal(cluster.metadata.allHosts().iterator(), true));
     }
 
@@ -160,8 +161,7 @@ class ControlConnection implements Host.StateListener {
         Connection connection = connectionRef.get();
         if (connection != null && connection.isDefunct()) {
             // If the connection was marked as defunct, this already reported the
-            // node down, which will trigger a reconnect. Otherwise, just reconnect
-            // manually
+            // node down, which will trigger a reconnect.
             return;
         }
         // If the connection is not defunct, or the host has left, just reconnect manually