+++ driver-core/src/main/java/com/datastax/driver/core/DynamicConnectionPool.java
@@ -188,6 +188,9 @@ class DynamicConnectionPool extends HostConnectionPool {
     }
 
     private PooledConnection waitForConnection(long timeout, TimeUnit unit) throws ConnectionException, TimeoutException {
+        if (timeout == 0)
+            throw new TimeoutException();
+
         long start = System.nanoTime();
         long remaining = timeout;
         do {
@@ -252,11 +255,21 @@ class DynamicConnectionPool extends HostConnectionPool {
                 close(connection);
         } else {
             if (connections.size() > options().getCoreConnectionsPerHost(hostDistance) && inFlight <= options().getMinSimultaneousRequestsPerConnectionThreshold(hostDistance)) {
-                trashConnection(connection);
+                connection.setTrashTimeIn(options().getIdleTimeoutSeconds());
             } else if (connection.maxAvailableStreams() < MIN_AVAILABLE_STREAMS) {
                 replaceConnection(connection);
             } else {
+                connection.cancelTrashTime();
                 signalAvailableConnection();
+            }
+        }
+    }
+
+    @Override
+    void trashIdleConnections(long now) {
+        for (PooledConnection connection : connections) {
+            if (connection.getTrashTime() < now) {
+                trashConnection(connection);
             }
         }
     }
@@ -277,6 +290,7 @@ class DynamicConnectionPool extends HostConnectionPool {
                 int opened = open.get();
                 if (opened <= options().getCoreConnectionsPerHost(hostDistance)) {
                     connection.markForTrash.set(false);
+                    connection.cancelTrashTime();
                     return false;
                 }
 
@@ -286,13 +300,9 @@ class DynamicConnectionPool extends HostConnectionPool {
 
             doTrashConnection(connection);
         }
-        // If compareAndSet failed, it means we raced and another thread will execute doTrashConnection.
-        // If the connection needs to be closed (inFlight == 0), we don't need to do it here because the other thread will necessarily do
-        // it:
-        // - the current thread decremented inFlight in returnConnection
-        // - we know it did it before the connection was trashed, because otherwise it would have entered `if (trash.contains(connection))`
-        //   in returnConnection and not arrived here.
-        // - so the other thread will see the up-to-date value of inFlight and take appropriate action.
+        // If compareAndSet failed, it means we raced with another thread that will execute doTrashConnection.
+        // Since trashConnection is called from a scheduled task, we're sure that the current thread did not modify
+        // inFlight, so the other thread will take care of closing the connection if necessary (i.e. if inFlight == 0).
         return true;
     }