+++ driver-core/src/main/java/com/datastax/driver/core/HostConnectionPool.java
@@ -59,6 +59,8 @@ abstract class HostConnectionPool {
 
     abstract void replaceDefunctConnection(final PooledConnection connection);
 
+    abstract void trashIdleConnections(long now);
+
     abstract int opened();
 
     abstract int inFlightQueriesCount();