+++ driver-core/src/main/java/com/datastax/driver/core/RequestHandler.java
@@ -57,6 +57,12 @@ class RequestHandler implements Connection.ResponseCallback {
     private volatile HostConnectionPool currentPool;
     private final AtomicReference<QueryState> queryStateRef;
 
+    // This represents the number of times a retry has been triggered by the RetryPolicy (this is different from
+    // queryStateRef.get().retryCount, because some retries don't involve the policy, for example after an
+    // OVERLOADED error).
+    // This is incremented by one writer at a time, so volatile is good enough.
+    private volatile int retriesByPolicy;
+
     private volatile ConsistencyLevel retryConsistencyLevel;
 
     private volatile Map<InetSocketAddress, Throwable> errors;
@@ -113,9 +119,7 @@ class RequestHandler implements Connection.ResponseCallback {
 
         PooledConnection connection = null;
         try {
-            // Note: this is not perfectly correct to use getConnectTimeoutMillis(), but
-            // until we provide a more fancy to control query timeouts, it's not a bad solution either
-            connection = currentPool.borrowConnection(manager.configuration().getSocketOptions().getConnectTimeoutMillis(), TimeUnit.MILLISECONDS);
+            connection = currentPool.borrowConnection(manager.configuration().getPoolingOptions().getPoolTimeoutMillis(), TimeUnit.MILLISECONDS);
             if (current != null) {
                 if (triedHosts == null)
                     triedHosts = new ArrayList<Host>();
@@ -134,8 +138,7 @@ class RequestHandler implements Connection.ResponseCallback {
             return false;
         } catch (BusyConnectionException e) {
             // The pool shouldn't have give us a busy connection unless we've maxed up the pool, so move on to the next host.
-            if (connection != null)
-                connection.release();
+            connection.release();
             logError(host.getSocketAddress(), e);
             return false;
         } catch (TimeoutException e) {
@@ -295,7 +298,7 @@ class RequestHandler implements Connection.ResponseCallback {
                                                               rte.getRequiredAcknowledgements(),
                                                               rte.getReceivedAcknowledgements(),
                                                               rte.wasDataRetrieved(),
-                                                              retryCount);
+                                                              retriesByPolicy);
 
                             if (metricsEnabled()) {
                                 if (retry.getType() == Type.RETRY)
@@ -315,7 +318,7 @@ class RequestHandler implements Connection.ResponseCallback {
                                                                wte.getWriteType(),
                                                                wte.getRequiredAcknowledgements(),
                                                                wte.getReceivedAcknowledgements(),
-                                                               retryCount);
+                                                               retriesByPolicy);
 
                             if (metricsEnabled()) {
                                 if (retry.getType() == Type.RETRY)
@@ -334,7 +337,7 @@ class RequestHandler implements Connection.ResponseCallback {
                                                               ue.getConsistencyLevel(),
                                                               ue.getRequiredReplicas(),
                                                               ue.getAliveReplicas(),
-                                                              retryCount);
+                                                              retriesByPolicy);
 
                             if (metricsEnabled()) {
                                 if (retry.getType() == Type.RETRY)
@@ -419,8 +422,9 @@ class RequestHandler implements Connection.ResponseCallback {
                     else {
                         switch (retry.getType()) {
                             case RETRY:
+                                ++retriesByPolicy;
                                 if (logger.isDebugEnabled())
-                                    logger.debug("Doing retry {} for query {} at consistency {}", retryCount, statement, retry.getRetryConsistencyLevel());
+                                    logger.debug("Doing retry {} for query {} at consistency {}", retriesByPolicy, statement, retry.getRetryConsistencyLevel());
                                 if (metricsEnabled())
                                     metrics().getErrorMetrics().getRetries().inc();
                                 retry(true, retry.getRetryConsistencyLevel());